+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.0.2] - 2023-7-1
+
+### Added
+
+- `enabled` variable
+    - if `True`, prints
+    - if `False`, won't print
+
+### Changed
+
+- Completely changed how to the variable name is retrieved
+
 ## [0.0.1] - 2023-7-1
 
 ### Released